pyproject.toml
[tool.poetry]
name = "ralmo"
version = "0.1.0"
description = "Resource-Adaptive Language Model Orchestration — local-first, confidence-gated hybrid LLM inference"
authors = ["MH"]
license = "MIT"
readme = "README.md"
packages = [{include = "ralmo_core", from = "src"}, {include = "cli", from = "src"}]

[tool.poetry.scripts]
ralmo = "cli.run_engine:app"

[tool.poetry.dependencies]
python = "^3.11"
llama-cpp-python = ">=0.2.0"
typer = {extras = ["all"], version = ">=0.9.0"}
hydra-core = ">=1.3.0"
omegaconf = ">=2.3.0"
rich = ">=13.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.21.0"
mypy = ">=1.0.0"
ruff = ">=0.1.0"
black = ">=23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
